abstracted token generation functionality from auth controller
