+++ backend/controllers/auth.js
@@ -1,6 +1,7 @@
 import jwt from "jsonwebtoken";
 import User from "../models/user.js";
 import MailService from "../service/mail.js";
+import tokenService from "../service/token.js";
 import generateRandomBytes from "../utils/generateRandomBytes.js";
 
 const ONE_DAY_IN_MILLISEC = 24 * 60 * 60 * 1000;
@@ -10,7 +11,6 @@ const ONE_DAY_IN_MILLISEC = 24 * 60 * 60 * 1000;
 // @access Public
 export const signup = async (req, res) => {
   const { email, password, firstName, lastName } = req.body;
-  console.log(email, password, firstName, lastName);
   if (!firstName || !lastName || !email || !password)
     return res.status(400).json({ message: "Please fill in all fields" });
 
@@ -50,8 +50,14 @@ export const signin = async (req, res) => {
     const match = await foundUser.matchPassword(password);
     if (!match) return res.status(400).json({ message: "Invalid credentials" });
 
-    const accessToken = generateAccessToken(foundUser.email, foundUser._id);
-    const refreshToken = generateRefreshToken(foundUser.email, foundUser._id);
+    const accessToken = tokenService.generateAccessToken(
+      foundUser.email,
+      foundUser._id
+    );
+    const refreshToken = tokenService.generateRefreshToken(
+      foundUser.email,
+      foundUser._id
+    );
 
     foundUser.refreshToken = refreshToken;
     await foundUser.save();
@@ -99,7 +105,10 @@ export const verifyRefreshToken = async (req, res) => {
     const decoded = jwt.verify(refreshToken, process.env.REFRESH_TOKEN_SECRET);
     if (foundUser.id !== decoded.id) return res.sendStatus(403);
 
-    const accessToken = generateAccessToken(foundUser.email, foundUser.id);
+    const accessToken = tokenService.generateAccessToken(
+      foundUser.email,
+      foundUser.id
+    );
     res.status(201).json({ accessToken, isAdmin: foundUser.isAdmin });
   } catch (error) {
     console.log(error);
@@ -114,20 +123,8 @@ export const verify = async (req, res) => {
     user.isConfirmed = true;
     await user.save();
 
-    return res.redirect("http://localhost:3000");
+    return res.redirect("http://localhost:3000"); //TODO: update client url for production
   } catch (error) {
     console.log(error);
   }
-};
-
-const generateAccessToken = (email, id) => {
-  return jwt.sign({ email, id }, process.env.ACCESS_TOKEN_SECRET, {
-    expiresIn: "10s",
-  });
-};
-
-const generateRefreshToken = (email, id) => {
-  return jwt.sign({ email, id }, process.env.REFRESH_TOKEN_SECRET, {
-    expiresIn: "15s",
-  });
 };
